requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
langchain
sentence-transformers

